+++ src/Config/ConfigDefinitions.cpp
@@ -31,17 +31,17 @@ const float RETURN_ACCEL = 16000;
 
 // Timing Settings (milliseconds)
 namespace Timing {
-const float CLAMP_SETTLE_TIME = 150.0f;
-const float CLAMP_RELEASE_TIME = 450.0f;
-const float HOME_SETTLE_TIME = 100.0f;
-const float MOTION_SETTLE_TIME = 50.0f;
-const float ALIGN_SETTLE_TIME = 300.0f;
-const float SIGNAL_DURATION = 2000.0f;
-const float MOTOR_TIMEOUT = 5000.0f;
+const int CLAMP_SETTLE_TIME = 150;
+const int CLAMP_RELEASE_TIME = 450;
+const int HOME_SETTLE_TIME = 100;
+const int MOTION_SETTLE_TIME = 50;
+const int ALIGN_SETTLE_TIME = 300;
+const int SIGNAL_DURATION = 2000;
+const int MOTOR_TIMEOUT = 5000;
 // Alignment-specific timing (ms)
-const float ALIGNMENT_LEFT_CLAMP_EXTEND_MS = 50.0f;   // allow left clamp to extend
-const float ALIGNMENT_ALIGNMENT_MOVE_MS = 20.0f;      // allow alignment movement to complete
-const float ALIGNMENT_RIGHT_CLAMP_WAIT_MS = 150.0f;   // right clamp extension time
-const float ALIGNMENT_SHORT_SETTLE_MS = 100.0f;       // short settle/pause
-const float ALIGNMENT_LONG_SETTLE_MS = 200.0f;        // long settle/pause
+const int ALIGNMENT_LEFT_CLAMP_EXTEND_MS = 50;   // allow left clamp to extend
+const int ALIGNMENT_ALIGNMENT_MOVE_MS = 20;      // allow alignment movement to complete
+const int ALIGNMENT_RIGHT_CLAMP_WAIT_MS = 150;   // right clamp extension time
+const int ALIGNMENT_SHORT_SETTLE_MS = 100;       // short settle/pause
+const int ALIGNMENT_LONG_SETTLE_MS = 200;        // long settle/pause
 }  // namespace Timing 3
